Fix time form
