+++ components/Dashboard.tsx
@@ -8,25 +8,7 @@ import {
 } from "./ui/card";
 import { Calendar } from "@/components/ui/calendar";
 import { Separator } from "./ui/separator";
-import { formatDistanceToNow } from "date-fns";
-
-function formatDistanceDay(date: Date): string {
-  const oneDay = 1000 * 3600 * 24;
-  const distance = Date.now() - date.getTime();
-  if (distance < oneDay && distance >= 0) {
-    return "Today";
-  }
-  if (distance < 2 * oneDay && distance > 0) {
-    return "Yesterday";
-  }
-
-  if (distance > -1 * oneDay && distance < 0) {
-    return "Tomorrow";
-  }
-  let result = formatDistanceToNow(date, { addSuffix: true });
-  result = result[0].toUpperCase() + result.slice(1);
-  return result;
-}
+import { formatDistanceDay } from "@/helpers/time";
 
 function Dashboard() {
   const [date, setDate] = useState<Date | undefined>(new Date());
